pyproject.toml
[project]
name = "bd-s25-a1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=6.29.5",
    "neo4j>=5.28.1",
    "numpy>=2.2.3",
    "pandas>=2.2.3",
    "psycopg2-binary>=2.9.10",
    "pymongo>=4.11.2",
    "ruff>=0.9.9",
]

[tool.ruff]
extend-include = ["*.ipynb"]
lint.extend-select = ["I"]
fix=true
line-length = 90
